maj sup cascade oeuvre (probleme rechargement)

+++ Symfony/src/ModuleGestionBundle/Controller/OeuvreController.php
@@ -397,15 +397,19 @@ class OeuvreController extends Controller
             // On récupère l'id de l'oeuvre a supprimée
             $id = $request->get('id');
 
-            $em = $this->getDoctrine()->getManager();
-            // On fait une requête pour récupérer les infos de l'oeuvre
-            $oeuvre = $em->getRepository('ModuleGestionBundle:Oeuvre')->find($id);
+            if($id != ""){
+                $em = $this->getDoctrine()->getManager();
+                // On fait une requête pour récupérer les infos de l'oeuvre
+                $oeuvre = $em->getRepository('ModuleGestionBundle:Oeuvre')->find($id);
 
-            // Suppression de l'oeuvre sélectionnée
-            $em->remove($oeuvre);
-            $em->flush();
+                // Suppression de l'oeuvre sélectionnée
+                $em->remove($oeuvre);
+                $em->flush();
 
-            $message = "Suppression effectuée avec succès !";
+                $message = "Suppression effectuée avec succès !";
+            }else{
+                $message = "Erreur de suppression !";
+            }
 
             // Puis on le renvoie dans un tableau en Json
             return new JsonResponse(array('msg' => json_encode($message)));
